+++ build.gradle
@@ -25,7 +25,7 @@ buildscript {
 }
 
 ext {
-    jxBrowserVersion = '7.0' // The version of JxBrowser used in the examples.
+    jxBrowserVersion = '7.1' // The version of JxBrowser used in the examples.
     guavaVersion = '25.0-jre' // Some of the examples use Guava.
 }
